fix: 関数やクラスの命名を変更
